requirements.txt
opencv-python
numpy
pandas
scikit-learn
flask
flask_cors
pillow
